Refactor End-to-End tests to support different executable
