+++ scripts/end_to_end_test/end_to_end_test_dmlf.yaml
@@ -2,25 +2,10 @@
 
 ---
 test_description: >
-  DISABLED - Single standalone node can mine transactions which are then queried
-
-setup_conditions:
-  test_name: "basic_standalone"
-  number_of_nodes: 1
-  mining_nodes: [0]
-  max_test_time: 60
-
-steps:
-  - send_txs: { "name": "batch1", "amount": 10, "nodes": [0] }
-  - verify_txs: { "name": "batch1", "nodes": [0] }
-...
-
----
-test_description: >
   Remote Etch execution using Dmlf in a client-server setup
 
 setup_conditions:
-  instance_type: "DmlfEtchInstance"
+  test_type: "DmlfEtch"
   test_name: "dmlf_etch_client_server"
   number_of_nodes: 2
   node_connections: []
